2025-Python-Django-Redis-in-Action-10-Project-2-Phase-2/app/polls/schema.py
from typing import Dict

from ninja import Schema


class VoteSchema(Schema):
    option: str


class PollOutListSchema(Schema):
    id: int
    question: str


class CreatePollSchema(Schema):
    question: str
    text: Dict[str, str]


class PollOutSchema(Schema):
    id: int
    question: str
    text: Dict[str, str]


class ErrorSchema(Schema):
    error: str
