2025-Python-Django-Redis-in-Action-10-Project-2-Phase-2/requirements.txt
Django==5.2
psycopg[binary,async]==3.2.6
redis==5.2.1
uvicorn==0.30.1
django-ninja==1.4.1